chore(mailchimp): typescripted and moved archiveUserInAudience to mailchimp module
